Schema changes quiz crud change needed

+++ QuizVerse.Domain/DatabaseScripts/Functions/get_quiz_data_by_id.sql
@@ -18,7 +18,7 @@ RETURNS TABLE (
     name VARCHAR,
     "quizCategoryId" INT,
     description VARCHAR,
-    "totalTime" INT,
+    "totalTime" NUMERIC,
     "difficultyLevelId" INT,
     "totalQuestion" INT,
     "isPaid" BOOLEAN,
